implement disassociate
Implement Disassociate request #14
