+++ typings/datapyrse/messages/_relate.pyi
@@ -1,0 +1,18 @@
+from dataclasses import dataclass
+from datapyrse._entity_metadata import EntityMetadata as EntityMetadata, ManyToManyRelationshipMetadata as ManyToManyRelationshipMetadata, ManyToOneRelationshipMetadata as ManyToOneRelationshipMetadata, OneToManyRelationshipMetadata as OneToManyRelationshipMetadata, OrgMetadata as OrgMetadata
+from datapyrse._entity_reference import EntityReference as EntityReference
+from datapyrse._entity_reference_collection import EntityReferenceCollection as EntityReferenceCollection
+from logging import Logger
+
+@dataclass
+class RelateRequest:
+    primary_record: EntityReference
+    related_records: EntityReferenceCollection
+    org_metadata: OrgMetadata
+    relationship_name: str | None = ...
+    relationship_type: OneToManyRelationshipMetadata | ManyToManyRelationshipMetadata | ManyToOneRelationshipMetadata | None = ...
+    logger: Logger = ...
+    def __post_init__(self) -> None: ...
+    def validate_relationship_name(self, logger: Logger | None = None) -> OneToManyRelationshipMetadata | ManyToOneRelationshipMetadata | ManyToManyRelationshipMetadata | None: ...
+    def parse_relationship_name(self) -> tuple[str | None, OneToManyRelationshipMetadata | ManyToManyRelationshipMetadata | ManyToOneRelationshipMetadata | None]: ...
+    def __init__(self, primary_record, related_records, org_metadata, relationship_name=..., logger=...) -> None: ...
